VITIS-8127 Refactor SubCmdConfigure and use new configuration JSON
