+++ src/runtime_src/core/tools/xbutil2/xbutil.cpp
@@ -36,6 +36,9 @@ R"(
         "contents": [{
             "name": "examine",
             "contents": ["dynamic-regions", "electrical", "host", "mechanical", "memory", "pcie-info", "platform", "thermal"]
+        },{
+          "name": "configure",
+          "contents": ["host-mem", "p2p"]
         }]
     },{
         "name": "alveo",
@@ -69,7 +72,7 @@ int main( int argc, char** argv )
     subCommands.emplace_back(std::make_shared<  SubCmdExamine  >(false, false, false, configTree));
     subCommands.emplace_back(std::make_shared<  SubCmdProgram  >(false, false, false));
     subCommands.emplace_back(std::make_shared<    SubCmdReset  >(false, false, false));
-    subCommands.emplace_back(std::make_shared< SubCmdConfigure >(false, false, false));
+    subCommands.emplace_back(std::make_shared< SubCmdConfigure >(false, false, false, configTree));
 
     // Parse sub commands from json files
     populateSubCommandsFromJSON(subCommands, executable);
